add status code chip
